requirements.txt
pandas
scikit-learn
joblib
google-cloud-pipeline-components
kfp